+++ spec/cuprum/rails/serializers/json/properties_serializer_spec.rb
@@ -2,10 +2,10 @@
 
 require 'cuprum/rails/serializers/json/properties_serializer'
 
-require 'support/examples/serializers/json_serializer_examples'
+require 'support/examples/serializers/json/properties_serializer_examples'
 
 RSpec.describe Cuprum::Rails::Serializers::Json::PropertiesSerializer do
-  include Spec::Support::Examples::Serializers::JsonSerializerExamples
+  include Spec::Support::Examples::Serializers::Json::PropertiesSerializerExamples # rubocop:disable Metrics/LineLength
 
   subject(:serializer) { described_class.new }